--- android/app/src/main/res/layout/activity_main.xml
@@ -1,209 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="20dp"
-    tools:context="edu.gatech.cs2340.rajsjarfiles.spacetrader.views.MainActivity"
-    android:background="@drawable/main_header_selector">
-
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Embark on Your Journey."
-        android:textFontWeight="800"
-        android:textSize="32dp"
-        android:layout_marginBottom="32dp"
-        android:textColor="@android:color/white"
-        />
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Your name:"
-        android:textColor="@android:color/white"
-        />
-
-
-    <EditText
-        android:id="@+id/player_name_input"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="22dp"
-        android:textColor="@android:color/white"
-        />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Skill Points (Total 16 pts):"
-        android:layout_marginBottom="12dp"
-        android:textColor="@android:color/white"
-        />
-
-    <TextView
-        android:id="@+id/skill_point_indicator"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="You have 16 pts remaining"
-        android:layout_marginBottom="12dp"
-        android:textColor="@android:color/holo_red_light"
-
-        />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-
-        >
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:textColor="@android:color/white"
-            >
-            <TextView
-                android:layout_width="match_parent"
-                android:textAlignment="center"
-                android:layout_height="wrap_content"
-                android:text="Pilot"
-                android:textColor="@android:color/white"
-                />
-            <EditText
-                android:id="@+id/pilot_skill_input"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:inputType="number"
-                android:maxLength="2"
-                android:textColor="@android:color/white"
-                />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="22dp"
-
-            android:orientation="vertical">
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:text="Fighter"
-                android:textColor="@android:color/white"
-                />
-            <EditText
-                android:id="@+id/fighter_skill_input"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:inputType="number"
-                android:maxLength="2"
-                android:textColor="@android:color/white"
-                />
-        </LinearLayout>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginBottom="22dp"
-        >
-        <LinearLayout
-            android:layout_width="1dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-            <TextView
-                android:layout_width="match_parent"
-                android:textAlignment="center"
-                android:layout_height="wrap_content"
-                android:text="Trader"
-                android:textColor="@android:color/white"
-                />
-            <EditText
-                android:id="@+id/trader_skill_input"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:inputType="number"
-                android:maxLength="2"
-                android:textColor="@android:color/white"
-                />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="1dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-            <TextView
-                android:layout_width="match_parent"
-                android:textAlignment="center"
-                android:layout_height="wrap_content"
-                android:text="Engineer"
-                android:textColor="@android:color/white"
-
-                />
-            <EditText
-                android:id="@+id/engineer_skill_input"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="center"
-                android:inputType="number"
-                android:maxLength="2"
-                android:textColor="@android:color/white"
-                />
-        </LinearLayout>
-
-    </LinearLayout>
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Difficulty:"
-        android:layout_marginBottom="12dp"
-        android:textColor="@android:color/white"
-
-        />
-    <Spinner
-        android:id="@+id/difficulty_spinner"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:spinnerMode="dropdown"
-        android:layout_marginBottom="22dp"
-        android:textColor="@android:color/white"
-
-        />
-    <Button
-        android:id="@+id/submit_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="22dp"
-        android:text="Let's Go"
-
-        android:radius="5dp"
-        android:textColor="@android:color/white"
-        android:background="#17EAD9"
-        android:onClick="onSave"/>
-
-    <Button
-        android:id="@+id/exit_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Exit"
-
-        android:radius="5dp"
-        android:textColor="@android:color/white"
-        android:background="#130CB7"
-        android:onClick="onExit"/>
-</LinearLayout>
