+++ android/app/src/main/res/values/strings.xml
@@ -1,8 +1,7 @@
 <resources>
     <string name="app_name">SpaceTrader</string>
-    <string name="title_activity_main">MainActivity</string>
-    <string name="title_activity_game">GameActivity</string>
     <string name="title_activity_start">StartActivity</string>
+    <string name="title_activity_game">GameActivity</string>
     <string name="title_activity_travel">TravelActivity</string>
     <string name="title_activity_market">MarketActivity</string>
     <string name="title_activity_ship">ShipActivity</string>
